+++ jormungandr/src/blockchain/process.rs
@@ -7,7 +7,7 @@ use super::{
 use crate::{
     blockcfg::{Block, Header, HeaderHash},
     blockchain::Checkpoints,
-    intercom::{self, BlockMsg, ExplorerMsg, NetworkMsg, PropagateMsg, TransactionMsg, WatchMsg},
+    intercom::{self, BlockMsg, NetworkMsg, PropagateMsg, TransactionMsg, WatchMsg},
     metrics::{Metrics, MetricsBackend},
     topology::NodeId,
     utils::{
@@ -56,7 +56,6 @@ pub struct TaskData {
     pub stats_counter: Metrics,
     pub network_msgbox: MessageBox<NetworkMsg>,
     pub fragment_msgbox: MessageBox<TransactionMsg>,
-    pub explorer_msgbox: Option<MessageBox<ExplorerMsg>>,
     pub watch_msgbox: MessageBox<WatchMsg>,
     pub garbage_collection_interval: Duration,
 }
@@ -74,7 +73,6 @@ struct Process {
     stats_counter: Metrics,
     network_msgbox: MessageBox<NetworkMsg>,
     fragment_msgbox: MessageBox<TransactionMsg>,
-    explorer_msgbox: Option<MessageBox<ExplorerMsg>>,
     watch_msgbox: MessageBox<WatchMsg>,
     garbage_collection_interval: Duration,
     tip_update_mbox: MessageBox<Arc<Ref>>,
@@ -145,7 +143,6 @@ pub async fn start(
         stats_counter,
         network_msgbox,
         fragment_msgbox,
-        explorer_msgbox,
         garbage_collection_interval,
         watch_msgbox,
     } = task_data;
@@ -162,7 +159,6 @@ pub async fn start(
         stats_counter,
         network_msgbox,
         fragment_msgbox,
-        explorer_msgbox,
         watch_msgbox,
         garbage_collection_interval,
         tip_update_mbox,
@@ -186,7 +182,6 @@ impl Process {
             self.blockchain_tip.clone(),
             self.blockchain.clone(),
             Some(self.fragment_msgbox.clone()),
-            self.explorer_msgbox.clone(),
             Some(self.watch_msgbox.clone()),
             self.stats_counter.clone(),
         );
@@ -204,7 +199,6 @@ impl Process {
         let blockchain = self.blockchain.clone();
         let blockchain_tip = self.blockchain_tip.clone();
         let network_msg_box = self.network_msgbox.clone();
-        let explorer_msg_box = self.explorer_msgbox.clone();
         let watch_msg_box = self.watch_msgbox.clone();
         let stats_counter = self.stats_counter.clone();
         tracing::trace!("handling new blockchain task item");
@@ -228,7 +222,6 @@ impl Process {
                         blockchain,
                         self.tip_update_mbox.clone(),
                         network_msg_box,
-                        explorer_msg_box,
                         watch_msg_box,
                         *leadership_block,
                     )
@@ -278,7 +271,6 @@ impl Process {
                         self.blockchain.clone(),
                         self.tip_update_mbox.clone(),
                         network_msg_box,
-                        explorer_msg_box,
                         watch_msg_box,
                         self.get_next_block_scheduler.clone(),
                         handle,
@@ -364,7 +356,6 @@ async fn process_leadership_block(
     mut blockchain: Blockchain,
     tip_update_mbox: MessageBox<Arc<Ref>>,
     network_msg_box: MessageBox<NetworkMsg>,
-    explorer_msg_box: Option<MessageBox<ExplorerMsg>>,
     mut watch_msg_box: MessageBox<WatchMsg>,
     leadership_block: LeadershipBlock,
 ) -> chain::Result<()> {
@@ -374,10 +365,6 @@ async fn process_leadership_block(
     watch_msg_box
         .send(WatchMsg::NewBlock(block.clone()))
         .await?;
-
-    if let Some(mut msg_box) = explorer_msg_box {
-        msg_box.send(ExplorerMsg::NewBlock(Box::new(block))).await?;
-    }
 
     process_and_propagate_new_ref(Arc::clone(&new_block_ref), tip_update_mbox, network_msg_box)
         .await?;
@@ -450,7 +437,6 @@ async fn process_network_blocks(
     blockchain: Blockchain,
     tip_update_mbox: MessageBox<Arc<Ref>>,
     network_msg_box: MessageBox<NetworkMsg>,
-    mut explorer_msg_box: Option<MessageBox<ExplorerMsg>>,
     mut watch_msg_box: MessageBox<WatchMsg>,
     mut get_next_block_scheduler: GetNextBlockScheduler,
     handle: intercom::RequestStreamHandle<Block, ()>,
@@ -466,7 +452,6 @@ async fn process_network_blocks(
                 let res = process_network_block(
                     &blockchain,
                     block.clone(),
-                    explorer_msg_box.as_mut(),
                     &mut watch_msg_box,
                     &mut get_next_block_scheduler,
                 )
@@ -515,7 +500,6 @@ async fn process_network_blocks(
 async fn process_network_block(
     blockchain: &Blockchain,
     block: Block,
-    explorer_msg_box: Option<&mut MessageBox<ExplorerMsg>>,
     watch_msg_box: &mut MessageBox<WatchMsg>,
     get_next_block_scheduler: &mut GetNextBlockScheduler,
 ) -> Result<Option<Arc<Ref>>, chain::Error> {
@@ -546,14 +530,7 @@ async fn process_network_block(
                 Err(Error::MissingParentBlock(parent_hash))
             }
             PreCheckedHeader::HeaderWithCache { parent_ref, .. } => {
-                let r = check_and_apply_block(
-                    blockchain,
-                    parent_ref,
-                    block,
-                    explorer_msg_box,
-                    watch_msg_box,
-                )
-                .await;
+                let r = check_and_apply_block(blockchain, parent_ref, block, watch_msg_box).await;
                 r
             }
         }
@@ -566,10 +543,8 @@ async fn check_and_apply_block(
     blockchain: &Blockchain,
     parent_ref: Arc<Ref>,
     block: Block,
-    explorer_msg_box: Option<&mut MessageBox<ExplorerMsg>>,
     watch_msg_box: &mut MessageBox<WatchMsg>,
 ) -> Result<Option<Arc<Ref>>, chain::Error> {
-    let explorer_enabled = explorer_msg_box.is_some();
     let post_checked = blockchain
         .post_check_header(
             block.header().clone(),
@@ -578,24 +553,14 @@ async fn check_and_apply_block(
         )
         .await?;
     tracing::debug!("applying block to storage");
-    let mut block_for_explorer = if explorer_enabled {
-        Some(block.clone())
-    } else {
-        None
-    };
+
     let block_for_watchers = block.clone();
+
     let applied_block = blockchain
         .apply_and_store_block(post_checked, block)
         .await?;
     if let AppliedBlock::New(block_ref) = applied_block {
         tracing::debug!("applied block to storage");
-        if let Some(msg_box) = explorer_msg_box {
-            msg_box
-                .try_send(ExplorerMsg::NewBlock(Box::new(
-                    block_for_explorer.take().unwrap(),
-                )))
-                .unwrap_or_else(|err| tracing::error!("cannot add block to explorer: {}", err));
-        }
 
         watch_msg_box
             .try_send(WatchMsg::NewBlock(block_for_watchers))