+++ testing/jormungandr-integration-tests/src/jcli/rest/vote.rs
@@ -42,7 +42,6 @@ pub fn test_correct_proposal_number_is_returned() {
         .with_committees(&[alice.to_committee_id()])
         .with_slots_per_epoch(60)
         .with_certs(vec![vote_plan_cert])
-        .with_explorer()
         .with_treasury(1_000.into())
         .build(&temp_dir);