+++ testing/jormungandr-integration-tests/src/jormungandr/fragments.rs
@@ -25,7 +25,6 @@ pub fn send_all_fragments() {
             .with_consensus_genesis_praos_active_slot_coeff(ActiveSlotCoefficient::MAXIMUM)
             .with_slot_duration(3)
             .with_linear_fees(LinearFee::new(1, 1, 1))
-            .with_explorer()
             .with_mempool(Mempool {
                 pool_max_entries: 1_000_000usize.into(),
                 log_max_entries: 1_000_000usize.into(),
@@ -64,6 +63,7 @@ pub fn send_all_fragments() {
     );
 
     fragment_generator.prepare(BlockDate::new(1, 0));
+
     time::wait_for_epoch(2, jormungandr.rest());
 
     let mem_checks: Vec<MemPoolCheck> = fragment_generator.send_all().unwrap();