+++ testing/jormungandr-integration-tests/src/jormungandr/genesis/rewards.rs
@@ -28,8 +28,7 @@ pub fn collect_reward() {
             .with_slots_per_epoch(20)
             .with_consensus_genesis_praos_active_slot_coeff(ActiveSlotCoefficient::MAXIMUM)
             .with_slot_duration(3)
-            .with_total_rewards_supply(1_000_000.into())
-            .with_explorer(),
+            .with_total_rewards_supply(1_000_000.into()),
     )
     .unwrap();