+++ testing/jormungandr-integration-tests/src/jormungandr/mempool.rs
@@ -49,7 +49,7 @@ pub fn dump_send_correct_fragments() {
         ConfigurationBuilder::new()
             .with_slots_per_epoch(60)
             .with_block_content_max_size(100000.into())
-            .with_explorer()
+            .with_slot_duration(1)
             .with_mempool(Mempool {
                 pool_max_entries: 1_000_000usize.into(),
                 log_max_entries: 1_000_000usize.into(),
@@ -114,7 +114,6 @@ pub fn dump_send_invalid_fragments() {
         vec![&sender, &receiver],
         ConfigurationBuilder::new()
             .with_slots_per_epoch(60)
-            .with_explorer()
             .with_mempool(Mempool {
                 pool_max_entries: 1_000_000usize.into(),
                 log_max_entries: 1_000_000usize.into(),
@@ -162,7 +161,7 @@ pub fn non_existing_folder() {
         vec![&sender, &receiver],
         ConfigurationBuilder::new()
             .with_slots_per_epoch(60)
-            .with_explorer()
+            .with_slot_duration(1)
             .with_mempool(Mempool {
                 pool_max_entries: 1_000_000usize.into(),
                 log_max_entries: 1_000_000usize.into(),
@@ -216,7 +215,6 @@ pub fn fragment_which_reached_mempool_should_be_persisted() {
         ConfigurationBuilder::new()
             .with_slots_per_epoch(60)
             .with_slot_duration(3)
-            .with_explorer()
             .with_mempool(Mempool {
                 pool_max_entries: 1usize.into(),
                 log_max_entries: 1000usize.into(),
@@ -256,7 +254,6 @@ pub fn fragment_which_is_not_in_fragment_log_should_be_persisted() {
         ConfigurationBuilder::new()
             .with_slots_per_epoch(60)
             .with_slot_duration(3)
-            .with_explorer()
             .with_mempool(Mempool {
                 pool_max_entries: 1000usize.into(),
                 log_max_entries: 1usize.into(),
@@ -296,7 +293,6 @@ pub fn pending_fragment_should_be_persisted() {
         ConfigurationBuilder::new()
             .with_slots_per_epoch(5)
             .with_slot_duration(60)
-            .with_explorer()
             .with_mempool(Mempool {
                 pool_max_entries: 10usize.into(),
                 log_max_entries: 10usize.into(),
@@ -341,7 +337,6 @@ pub fn node_should_pickup_log_after_restart() {
     let config = ConfigurationBuilder::new()
         .with_slots_per_epoch(60)
         .with_slot_duration(3)
-        .with_explorer()
         .with_mempool(Mempool {
             pool_max_entries: 1usize.into(),
             log_max_entries: 1000usize.into(),