+++ testing/jormungandr-integration-tests/src/networking/explorer.rs
@@ -6,7 +6,6 @@ use hersir::builder::SpawnParams;
 use hersir::builder::Topology;
 use jormungandr_automation::testing::time;
 use jormungandr_lib::interfaces::BlockDate;
-use jormungandr_lib::interfaces::Explorer;
 use thor::FragmentSender;
 const LEADER_1: &str = "Leader_1";
 const LEADER_2: &str = "Leader_2";
@@ -70,12 +69,7 @@ pub fn passive_node_explorer() {
         .unwrap();
 
     let passive = controller
-        .spawn(
-            SpawnParams::new(PASSIVE)
-                .passive()
-                .in_memory()
-                .explorer(Explorer { enabled: true }),
-        )
+        .spawn(SpawnParams::new(PASSIVE).passive().in_memory())
         .unwrap();
     let mut alice = controller.wallet(ALICE).unwrap();
     let bob = controller.wallet(BOB).unwrap();