+++ testing/jormungandr-integration-tests/src/non_functional/voting/public/mod.rs
@@ -72,7 +72,6 @@ pub fn public_vote_load_scenario(quick_config: PublicVotingLoadTestConfig) {
         .with_committees(&[committee.to_committee_id()])
         .with_slots_per_epoch(quick_config.slots_in_epoch())
         .with_certs(vec![vote_plan_cert])
-        .with_explorer()
         .with_slot_duration(quick_config.slot_duration())
         .with_treasury(1_000.into())
         .build(&temp_dir);
@@ -203,7 +202,6 @@ pub fn adversary_public_vote_load_scenario(
         .with_committees(&[committee.to_committee_id()])
         .with_slots_per_epoch(quick_config.slots_in_epoch())
         .with_certs(vec![vote_plan_cert])
-        .with_explorer()
         .with_slot_duration(quick_config.slot_duration())
         .with_block_content_max_size(quick_config.block_content_max_size().into())
         .with_treasury(1_000.into())