+++ CHANGELOG.md
@@ -22,7 +22,7 @@
 - Change `blockContentSizeAvg` to represent information as a percentage of the block max size.
 - Bump ed25519-bip32 from 0.4.0 to 0.4.1
 - Now the tally is incremental and is always available in the rest API. The
-  EncryptedVoteTally certificate serves no purpose anymore and is thus removed.
+- Add standalone explorer crate.
 
 ## Release 0.13.0